fix: money keeps its cents and totals convert each record from its own currency
- DecimalInputFormatter (2 decimals, comma to dot) on every price/cost field
- Expense.amount is a double, so cents survive Firestore; older whole-number docs still load
- fuel records save and load their currency
- statistics, chart, history and analytics sum records via CurrencyService.toUah and wait for the rates
- tests: formatter, per-currency monthly total, cents/currency round trip

+++ lib/features/expenses/data/models/car_wash_record.dart
@@ -84,7 +84,7 @@ class CarWashRecord extends BaseRecord {
     return Expense(
       id: id,
       date: date,
-      amount: amount.toInt(),
+      amount: amount,
       category: ExpenseCategory.carWash,
       mileage: mileage,
       comment: comment ?? "Car Wash",
